Added my own robust validation
I included the original files along with the ones I modified to show
where I diverged. The assignment was to add some validation to what was
already provided but to achieve a certain look, I changed quite a bit
more, including changing the names of the files.  The txt lists the
changes I made. This version does NOT have regular expressions in it.